% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{inauguration_palette}
\alias{inauguration_palette}
\title{Inauguration palette generator}
\usage{
inauguration_palette(name, n)
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{inauguration_2021}
@importFrom graphics rgb rect par image text}

\item{n}{Number of colors desired. Most palettes now only
have 5 to 8 colors. Most color
schemes are derived from \href{http://poppygall.com/blog/2011/07/25/color-inspiration-national-park-wpa-posters/}{Color Inspiration – National Park WPA Posters}.
If omitted, uses all colours.}
}
\value{
A vector of colours.
}
\description{
Inauguration palette generator
}
\examples{
inauguration_2021_palette("inauguration_2021")
inauguration_2021("inauguration_2021", 3)
}
\keyword{colors}
