% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{inauguration_palettes}
\alias{inauguration_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 1.
}
\usage{
inauguration_palettes
}
\description{
Use \code{\link{inauguration}} to construct palettes of desired length.
}
\keyword{datasets}
